feat: 검증된 Dealer 모델 기능 구현
