+++ README.md
@@ -43,7 +43,7 @@
     - [x] 이름은 중복 될 수 없다.
 - 딜러
   - 딜러는 이름과 블랙잭을 가지고 있다.
-    - [ ] 딜러의 이름은 생성자로 구현된다. (딜러)
+    - [x] 딜러의 이름은 생성자로 구현된다. (딜러)
 - 참여자
   - [x] 참여자는 딜러 혹은 플레이어가 된다.
 - 참여자들
