+++ src/test/kotlin/blackjack/model/DealerTest.kt
@@ -6,18 +6,6 @@ import org.junit.jupiter.api.assertThrows
 
 class DealerTest {
 
-    class Dealer(name: String = DEFAULT_DEALER_NAME) : Participant(name) {
-
-        fun openFirstCard(): Card {
-            return getCards().firstOrNull() ?: throw IllegalArgumentException(ERROR_CARD_INDEX)
-        }
-
-        companion object {
-            private const val DEFAULT_DEALER_NAME = "딜러"
-            private const val ERROR_CARD_INDEX = "가지고 있는 카드가 없습니다."
-        }
-    }
-
     @Test
     fun `딜러는 첫번째 카드를 공개할 수 있다`() {
         val dealer = Dealer()
